Added comments on comparing my files to the solution files
Added comments to the index.html and style.css files after comparing my files to the solution files for this assignment. Only big mistake was not adding a class selector in addition to a id selector in the html file that would of saved me from writing redundant code in the css file.