My attempt at solving for the 04 excersise.
The differences between my changes and the solutions are as follows:

for the html containers I followed the general structure in the solution with the exception of indiviually naming each "info" class selector as whatever plant that container was for.

As for the CSS I actually put alot of the flexbox descriptions in the body type selector which I realize maybe isnt the best move if you want other parts of your page to have different display modes, but worked for this this task. This also led me to not add a bottom margin to the .title class selector but instead just add a gap of 32px in the body type selector. my ".container" for the 4 "items" was labeled ".plants. It did not contain the justify-content: center; description as the body once again housed that but did have the display:flex and a gap: 52px description. for the for item class selectors I just used a comma and listed them unlike the solution which just has them all labeled the same thing. They have a max-width: 200px; description, while I instead opted to put the flex: 0 0 200px in the .plants selector, which I think achieves the same thing??