My attempt at completeing 05 assignment for flexbox
The HTML structure was setup exaclty like the solution so yah!

For the .modal selector I did not add the gap: 16px; description but the rest was added.

I did add the flex-shrink: 0; descriptionfor the .icon class selector.

For the .header class selector I did not add the align-items: center; and did not set a bottom margin. When trying align-items: center; it did not seem to change what I had so not sure whats standing in for that lol but it works.

for the .text class selector I actually set padding on the top and bottom instead of just a bottom margin this to also compensate for the lack of margin in the .header class selector.