My first attempt at styling for the 06 flex assignment
This is before I have had the chance to comapre to the solutions so not much to comment other than I really like the end result in comparison to the desired-outcome picture, and have started making use of an online resource that can give exact color values from uploaded picture in an effort to better match colors from the picture.

I'd say as well that I checked off all of the self-check bullets for what I should of input for the css file for this assignment.